# docs/Mobile_App_API/OpenAPI_Swagger.yaml
openapi: 3.0.3
info:
  title: HRMS Mobile App API
  version: 1.0.0
  description: >-
    Employee mobile app API. OTP login linked to Temporary Employee records; 7-step resume-safe registration;
    approval-gated dashboard. Every response uses { success, message, data, errors, timestamp }.
servers:
  - url: http://localhost:5000
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/app/attendance:
    get:
      tags:
        - Dashboard
      summary: Attendance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              example:
                success: true
                message: Success
                data:
                  summary:
                    totalRecords: 22
                    presentDays: 20
                  records:
                    - date: '2026-06-24'
                      status: Present
                      clockIn: '09:05'
                      clockOut: '18:10'
                      hoursWorked: 8.5
                errors: []
                timestamp: '2026-06-25T12:00:00.000Z'
        '403':
          description: NOT_APPROVED
          content:
            application/json:
              example:
                success: false
                message: Dashboard is available only after HR approval.
                data: null
                errors:
                  - code: NOT_APPROVED
                    message: Dashboard is available only after HR approval.
                timestamp: '2026-06-25T12:00:00.000Z'
...
  /api/app/payroll:
    get:
      tags:
        - Dashboard
      summary: Payroll
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              example:
                success: true
                message: Success
                data:
                  payslips:
                    - id: 5
                      month: June
                      year: 2026
                      basicSalary: 21000
                      allowances: 15000
                      deductions: 4000
                      netSalary: 42000
                      paymentStatus: pending
                      payslipGenerated: false
                errors: []
                timestamp: '2026-06-25T12:00:00.000Z'
        '403':
          description: NOT_APPROVED
          content:
            application/json:
              example:
                success: false
                message: Dashboard is available only after HR approval.
                data: null
                errors:
                  - code: NOT_APPROVED
                    message: Dashboard is available only after HR approval.
                timestamp: '2026-06-25T12:00:00.000Z'
  /api/app/notifications:
    get:
      tags:
        - Dashboard
      summary: Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              example:
                success: true
                message: Success
                data:
                  notifications:
                    - id: 9
                      type: announcement
                      title: Holiday
                      message: Office closed on Friday
                      read: false
                      timestamp: 2026-06-24 10:00
                errors: []
                timestamp: '2026-06-25T12:00:00.000Z'
        '403':
          description: NOT_APPROVED
          content:
            application/json:
              example:
                success: false
                message: Dashboard is available only after HR approval.
                data: null
                errors:
                  - code: NOT_APPROVED
                    message: Dashboard is available only after HR approval.
                timestamp: '2026-06-25T12:00:00.000Z'
...